Fix for a bug which was exposed by ASan as well as no less than three further bugs exposed by the work to first the first.
The first bug was that the quit thread was passed a reference to two temporary weak_ptr which have been destroyed in the main thread. The fix was to pass them by value. Incidentally, the main thread code was tweaked to make the construction of the weak_ptr<>s from shared_ptr<>s an explicit operation. To test this, the first thought was to add a command to the broker which would cause it to issue a signal to itself, but this was felt to be too unlike the real usage (where the signal comes from another process). However, this change of heart came after making the command processing code table driven and mockable and having a unit test for its basic functionality.

The second attempt at a test uses a python system test to start the broken, run the broker-poker to ensure the broker is initialized, and the uses the python runtime to generate the signal.

Sadly, but perhaps inevitably, this exposed some other problems. Lit likes to run the tests is parallel (which is a good thing). Unfortunately, it's intended that there's only a single broker process: this led to the multiple brokers arguing over the shared pipe. The solution to that was to allow the user to specify the pipe name on the command-line. Unfortunately, this then shows that using a bad name resulted in error messages, but didn't lead to the broker exiting. This is also now fixed.

Finally, these changes exposed a bug in the maybe<> type which didn't implement operator=() and lacked unit testing.