Fix a compilation failure with the use of constexpr in the uint128 code.
The uint128 class was built on std::pair<>. In libcxx the constexpr support for std::pair<> is (correctly) determined by the language standard selected: in C++11 it wasn't all constexpr, so the compilation failed. This change removes the use of std::pair (which was a convenience for the implementation of the relational operators). Since the implementation is no longer really trivial, it also adds some basic unit testing for the implementation.