Eliminate the use of optionparser from the pstore-index-structure utility.
This extends the cmd_util::cl implementation slightly to support the cl::Optional and cl::OneOrMore options being used by the LLVM switch code. This allows the two implementations to share the same code. Improve the help text to include a simple usage string and to eliminate positional and alias values from the list of options.