Fix a build-break on Windows. The index-structure tool was trying to pass wchar_t to the command-line parser which expected UTF-8 char*.
Add an extra (Windows-only) function to the command-line parser which takes UTF-16 wchar_t and converts it to UTF-8 for processing by the existing code. Remove various copies of this code that had sprung up in the other utility programs. We weren't enabling Unicode mode for the libraries (again, only on Windows), so changes the cmake add_pstore() function to do so.