Fix for a bug in the broker's win32 read loop.
It was possible for the win32 pipe reading code to initiate more than one simultaneous read but it only had a single instance of the OVERLAPPED structure. These are now maintained in a list of in-flight reads using the "intrusive list" type. Any remaining reads are cancelled when the read loop exits.

The broker-poker flood test was extended to force it to make multiple connections from multiple threads in order to increase the pressure on the server when under test. A command-line switch on the broker allows the number of read threads to be explicitly specified.