A step in fixing both <https://github.com/SNSystems/pstore/issues/4> and <https://github.com/SNSystems/llvm-prepo/issues/6>.
This change uses the new sstring_view class in the name set to improve the efficiency of string comparisons. Unfortunately, adding database serialization capability for this type turned the circular dependency between the hamt_map, db_archive, and transaction classes from an annoyance into a compile-time failure which I had to resolve. This was done by splitting the transaction class into two (transaction<> and transaction_base) where the bulk of the implementation is in transaction_base which can be safely moved to a .cpp file. I've also moved all of the index types (uint128, hamt_set<sstring_view>, hamt_map<uint128,record>, etc.) into a module of their own and placing the get_ and flush_index() code into that same module.