On macOS, switch from ASL to os_log() (https://github.com/SNSystems/pstore/issues/9). Remove ostream interface.
Apple deprecated the former API and it results in warnings from the compiler. This change (finally) eliminates those warnings. A second part of this change removes support for arbitrary arguments to the log() function which would then use iostreams to convert the text to a string. This was obviously somewhat inefficient and also rarely used in the code.