Complete work for <https://github.com/SNSystems/pstore/issues/10>. Add pstore-diff tool.
It would be very useful to be able to easily see the delta between two
pstore revisions. The tool should to generate a list of all of the keys
updated in each of the indices between two specified revisions (i.e. the
version visible in revision1 is different from that in revision2).
Following the style of pstore-dump, the output should use YAML.