Fix a crash that could occur if a json parser instance was moved or copied.
The parser holds instances of the matcher<> subclasses in a stack: this makes up the parse stack. Prior to this change each matcher<> object had a reference to its owning parser; this was used to enable matchers to communicate with the parser in order to create new matcher instances, set the error state, and so on. When the parser was moved or copied, these pointers could end up dangling. The code now passes the parser to each function, eliminating the need for that additional bit of state but adding some parameter-passing noise to the code.