Fix an assertion in parallel_for_each() when the work item count is slightly larger than the worker thread count.
parallel_for_each() tries to apportion equal amounts of work to each worker thread. If the number of work items is not a multiple of the number of threads then it will allocate less work to the last thread. If we have 8 workers and 9 work items then the choices are either to allocate the work as [2,2,2,2,1,0,0,0] or [1,1,1,1,1,1,1,2]. Although the latter division of work is possibly superior, the algorithm choses the former. This means that we may not allocate the full 8 threads. This fix corrects an assertion so that it allows this case.