Add quit handling in the status server.
Extend the quit handling so that it shuts down the status server in
addition to the existing threads. This is done by sending the JSON
message {"quit":true}. This means that the shutdown() function needs
to know whether the server is listening to a UD or INET socket.

Add NOMINMAX everywhere that Windows.h is included to limit the number
of times that we have to undef min and max. Move array_elements<> to a
file of its own and remove the duplicate copies.