Fixme fixed. A more logical home for mprotect behavior tests.
There are a couple of unit tests for the code that marks pages as read
only once a transaction is committed. These used to live alongside the
database unit tests, but that wasn't a good choice (hence the
FIXME). This change moves those to a new test_protect module.

In addition there were a number of copies of the (utterly trivial)
mock_mutex type and these are rationalized to one in mock_mutex.hpp.

Re-run the boilerplate tool.