Fix a fixme in the core database code: rename is now a member of file.
The file class has a path member which "knows" the name of the file that
it is managing. This was previously set by the open() member. Unlink
(delete) was a non-member function which accepted the from- and to-path
to be renamed. This meant that if an open file was renamed, the file
object's idea of the path would not match the reality.

With this change, the file path is passed to the file ctor and the unlink
function is a member of file.

Logging now manages a list of logging destinations than than just one.