Enable the status server to be shut down without external state.
Remove support for Unix domain sockets because they were making it
difficult to keep the code working correctly on Windows and
non-Windows hosts.

Previously, the quit thread shut the status server down by reading the
port number from the broker.status file. This was a problem if two
brokers were running (as happens in the system tests). The second
broker would replace the file written by the first and it became
impossible to cleanly shutdown the first.

We now have a small object which is used by the command process and
quit thread to know the port number to which the quit message is to be send.