Eliminate the 'source' type from the JSON parser.
The source type was originally intended to represent the source of
data for the parse. This would have had subclasses to read from file,
network, etc. This design was then changed to separate these two
concerns: the JSON parser is now called multiple times passing it a
span of characters are they are available. A call to parser.eof()
indicates that the input is complete.

The model of push() and pop() from the source object meant that the
parse was unable to accurately track line and column numbers because
of the CR/LF combination where knowing which part of the pair would
have added complexity to the source class.

This change eliminates all of this. A matcher returns an additional
bool indicating whether the input characteer should be consumed. This
puts the top-level loop in complete control of the characters handed
to the matchers decoupling them from the input span altogether.