Fix the broker build inside LLVM.
The broker uses exceptions, but LLVM does not. This means that when
built inside LLVM, most of pstore is built without
exceptions. Anything against which the broker links have have an
additional "exceptions enabled" ("-ex" suffix) build. Unfortunately,
several libraries were violating this rule: particularly we were
linking against pstore-json-lib which in turns links pstore-dump-lib
which then links in almost the whole world.

This change adds an exception-enabled build of pstore-json lib and
decouples it from the dump library. This severely limits what the
status_server can now currently do, but this will be addressed soon.

Most of the json unit tests need to be re-written due to the lack of
dump-lib.