Add size (member count) tracking to the index.
Previously, the size() member function was only used very rarey so we
didn't bother caching its result; instead it was computed using
std::distance(). However, I'm now anticipating that the function might
be used much more.

Adding a size_ member variable to the index is trivial, but this needs
to be persistent. To accomplish this, this change introduces a new
header block (containing signature, root pointer, and size) that is
written by flush() and read by the ctor.

Bump the database verion since this is incompatible with files built
by earlier versions.