Use allocate() rather than alloc_rw() for alignment. Eliminate duplicated check.
store_leaf_node() needs to ensure that it allocates storage that's 4-byte aligned: alloc_rw() was used to do this. alloc_rw() calls allocate() to make the space followed by getrw() to get an r/w pointer to it. However, we don't actually need or use the returned pointer so its simpler to just use allocate() directly.

The check that modifications are being make to an index from the database's active revision was repeated in several methods. These methods are all wrappers for insert_or_upsert(), so the check is relocated to there.