The names index now holds a collection of string addresses.
In order to try to cluster the string address as tightly as possible,
We now store a pointer to the string body in the names index rather
than the string itself. This means that code using the names index
needs to know about this extra indirection and code writing to the
index in particular needs to do so in two steps. The first stage adds
the string point and the second writes the new bodies (an
indirect_string_adder helper class deals with the ins and outs of this).