Remove the parser<>::input (czstring) API. Add an iterator-API.
The czstring API was removed because the code was producing a (fairly
minor) warning from Clang and because it was only used by the unit
test code. Instead add a version of the API which operates over an
iterator range. The full set of input() overrides is now: std::string,
gsl::span<>, and input iterators.