Replace the diff engine with a far more efficient implementation.
The previous implementation did the job, but wasn't terribly smart. It worked by building the set of key in an index at the "new" revision then removing all of those that are still found with the database synced to the "old" revision. The new implementation uses the data layout of the pstore file to optimize the process.

If first determines the address of the "old" transaction's footer beyond which everything is "new". It then traverses the new indexes. Any pointers -- be they to intermediate nodes or leaves -- which are older than the "old" threshold are immediately excluded from the traversal. Any remaining leaf nodes are added to the result.

The unused database::is_synced_to_head() method was removed.