Add Debug and Release builds to Travis CI.
The Travis YAML file is now generated by
utils/write_travis_yml.py. This is necessary because of the
duplication that results from the use of matrix:include to explicitly
define each of the builds that we need Travis to perform. This is in
turn because the standard matrix style build doesn't support the
addons->apt syntax (see
<https://github.com/travis-ci/travis-yaml/issues/58>). Revisit once
this bug is fixed.