Fix a warning from gcc7 about ignoring the result of strerror_r().
It turns out that there are two variants of strerror_r(). The POSIX
version returns an int and the GNU version  returns 'char *'. When
_GNU_SOURCE is defined, glibc provides the char * version. The GNU
function doesn't usually modify the provided buffer-- it only modifies
it sometimes. If you end up with the GNU version, you must use the
return value of the function as the string to print.