Fixes for the randomly failing the broker_kill system test.
Deal with the various status that the status server can be in:
initializing, listening, and closed. The server object can be
available (as managed by the shared_ptr/weak_ptr) but the server in
the process or initializing, or having already exited.

Add cygwin support in the broker_kill test.