Enable PSTORE_ALWAYS_SPANNING in Debug builds. Fix a HAMT unittest.
PSTORE_SPANNING mode causes the library to return every access to the
pstore in a freshly allocated shared_ptr<>. If, disabled, this happens
very rarely where a request spans two or more regions. This is a
debugging aid useful for shaking out a class of coding error.

Enabling it uncovered a bug in one of the HAMT unit tests. It was
loading an internal node with a single child and modifying two
elements of the child array.