Prevent code from getting a non-const pointer to committed data.
It was possible to use one of the getrw() functions to obtain a
non-const pointer to data in a committed transaction. This is bad
because, as an append-only store, we need to ensure that committed
data is immutable.

There were a number of unit generation_iterator unit tests that were
using this back door to deliberate corrup the store and verify
behavior. Since the access is no longer possible, the tests are removed.