CID305723: fragment::dependents() should not have been noexcept.
fragment::at<> could previously throw because it called
offset_to_section(). This latter function would throw if the section
alignment was not valid. This has now been tweaked so that
offset_to_section() simply asserts that the alignment is sufficient
and no longer throws. This enables at<> and dependents() to be noexcept.