CID264236: Fix for a TOCTOU vulnerability in the broker read pipe.
The code previously called stat() if mkfifo() failed with EEXIST to
find out if the file was a FIFO. An attacker could change the
filename's file association or other attributes between the check and
use.

Instead, we now try to open the file. If this succeed then, using the
same file descriptor, we check its attributes. In addition we do this
for both the read and the write end of the pipe: previously just the
read descriptor was checked.