Ensure only 1 thread tries to open the server pipe at a time.
Commit 3699a8355fd22be3c3fbefe2a4a8463bccb50844 fixed a TOCTOU problem
in which the process of opening the server pipe could be
attacked. Unfortunately, it introduced a race condition in which more
than one thread could try to call mkfifo(). If this happened, one of
them would be denied and the library will raise an error.

I now use a mutex to guarantee that one a single thread opens the
server pipe at a time.