Move fragment_content_iterator -> pointee_adaptor.
pstore::repo::details::fragment_content iterator was used by client code to access an array of pointers as if it were an array of objects. It shouldn't have been in the "details" namespace (the name "details" is intended to denote that its contents are an implementation detail and not for client use). It was also implementing a behaviour which isn't specific to fragments or their contents. For these, reasons I've moved this class to the pstore-support library and renamed it "pointee adaptor" (which seems to more closely reflect its behavior). The iterator category is now inherited from the underlying iterator and the full set of methods for a random-access-iterator has been added along with a set of unit tests.