Fix error handling for range_lock dtor and fix documentation.
The result of range_lock::lock() now returns the lock state (unless
already locked in which case the result remains as false). Merged the
implementations of lock() and try_lock(). Removed assertions from
lock(), unlock(), and the dtor in favor of defining sensible behavior
in some error cases. The dtor now unlocks the range rather than being
a no-op.