Add the debug-line section.
Debug-line section is an aggregation of an extent and a
generic-section. The latter is a little overkill in that it stores
alignment and xfixups which we don't need for debug-line, but reusing
it is expedient.

The overall DWARF .debug_line section is split such that the header is
stored separately and the portion of the line program that relates to
an individual function is stored as a debug_line section in the
function's fragment. The debug_line section then holds the extent of
that header.