Remove the databse member from hamt_map.
This is part of work to enable more correct use of const instances of
the database object. The first hurdle was to remove the non-const
database member of the index containers. This requires passing a
database to a number of the methods and adding a range<> type which
enables use of the container in range-based for loops.