Add explicit ctors for external_fixup.
The CI build is reporting a reference to uninitialized memory from
valgrind. <https://travis-ci.org/SNSystems/pstore/jobs/446105277>. I'm
unable to reproduce the failure locally, but have a theory that the
version of GCC in use by Travis isn't initializing the padding bytes
in the external_fixup structure. When gtest tries to dump the contents
of the struct, the valgrind check fires.