Change to silence CID 323822.
Coverity (correctly) spots that one of the switch cases is
unreachable. However, that case is only really there to silence
warnings from the compiler (and the associated statement is simply
"assert(false)").

To make both tools, the controlling for loop now deliberately executes
this code-path which becomes a no-op.

Change the indices enum to an enum class and added the required
type-casts.