Second try to fix a valgrind failure on CI.
The failure was caused by clang-3.8 not initializing the padding bytes
in the external_fixup type and then google test trying to hex-dump the
structure contents. The previous fix was almost, not not quite, right
in that it didn't include all of the padding that was being added by
the compiler.