Add support for native __uint128_t for compilers that support it.
Modified the constructors that accept and array of bytes so that in input is interpreted in strictly big-endian order. Added workarounds for a bug in VS 2017 aligned_storage<> type where the required alignment > alignof (max_align_t). Add more static_asserts for the ticket/compilation_member struct layout and alignment.