Avoid std::copy() for building payload in the generic_section ctor.
In the generic_section ctor, the memory beyond "this" (which is where
the three payload arrays live) is not initialized. That means that
using std::copy() is dangerous because it will assume that the
destination iterator points to valid instances which can safely use
the destination type's operator=().

Instead use placement new to build the internal/external fixup arrays
and memcpy() to the data blob.