Extent uint128 so that it can be used by sparse_array<>.
Add the various operators to the uint128 type so that it can be used
by sparse_array<> in place of one of the built-in integer types. Move
it to the support library. Extent the sparse_array<> tests so that it
uses a variety of bitmap types (rather than always using the default
uint64_t). Fix the resulting failures from this extended testing.