Changes for compatibility with the clang3.4 used by Klee.
Klee (http://klee.github.io) is currently built on a relatively old
version of clang (3.4): these changes are to enable the code to be
compiled. The intention is then to use Klee to improve testing for
some of the pstore APIs.

Add support for LLVM coverage if PSTORE_COVERAGE is defined.