Work to fix intermittent broker test failures from Win32 CI.
An example of these failures is AppVeyor build 1.0.151 in which the
broker appears to exit without cleanly shutting down the status server
and without completing main(). (There's a possibility that the output
is buffered and lost when the process exits, so this is largely
guesswork.)

Add some additional trace messages so that a better understanding can
be built if this failure occurs again.

A possibility is that the status server's attempt to delete its "port
file" results in an exception being thrown from a dtor and the process
is therefore terminated by the C++ runtime. Check all of the noexcept
decoration in this area.

The "pipe ended" error code is now down-graded to the debug trace
rather than being regarded as a true error. This makes the logs less
scary for the case in which the program is working correctly!