Fix disable_clang_warning_if_possible.
The CI clang 3.8 builds started issuing a lot of warnings that should
have been disabled by calling disable_clang_warning_if_possible with
-Wno-nullability-externtion.

I didn't realise that the check_c_source_compiles function caches the
result for the variable passed as its second argument. If the same
variable is passed then the result is retreived from the cache. This
defeated the attempt to discover whether a particular switch was
supported by the compiler. The fix is to use a different cmake
variable for each of the compiler switches that we're testing.