Revise the clang-tidy code so that it supports cmake 3.4.
I need to support cmake 3.4 because that's the latest version
currently supported by Solaris OpenCSW. The clang-tidy target code was
using a cmake feature which required 3.6 so this has been rewritten to
avoid that usage.

At the same time I've eliminated the clang-tidy target code that was
duplicated in a couple of the CMakeLists.txt files and moved it into
cmake/clang_tidy.cmake. That code has been revised so that it only
tests for the clang-tidy executable and the host compiler default
include paths once.

Adding clang-tody target is now simply a matter of calling
add_clang_tidy_target(). Eventually this could be folded into the
add_pstore... code so that it isn't a separate step but I'm not quite
ready to do that yet.

Modiy the readme file so that it documents the correct minimum cmake
version.