Add KLEE tests for the uint128 code along with resulting fixes.
THis commit adds a crude framework for performing testing using the
KLEE symbolic execution engine. Test code is run through klee and the
resulting test cases are then executed.

This exposed a couple of undefined-behaviour bugs in the uint128 code
when shift distance is 0. Added a compilation option when allows
support for the native __uint128_t type to be disabled even if the
compiler supports it. That then enables our pstore::uint128 type to be
compared to the compiler's native type.