Initial commit of the http server to be used for broker status.
A standalone executable (pstore-httpd) is a very simple web server
which server up the contents of the tools/httpd/html directory. This
path is converted to a ROM file-system at built-time and the resulting
structures become the data that the web server will return.

Note that this code currently compiles and but crashes on Windows.