Add private operator new/delete to generic_section.
The generic_section ctor assumes that memory has been allocated for
the variable length arrays that follow an instance of this type. This
is performed by the generic_section_creation_dispatcher. Prevent usage
of standard operator new and delete to avoid the possibility that
someone forgets this in the future and a memory overrun bug is
introduced.