Add an API to enable the server to be gracefully shutdown.
The URI instructing the server to exit has the form
/cmd/quit?magic=<random> where <random> is a 16 digit random string of
hex. The quit() function knows the correct value to use and sends the
URI to the server which responds by exiting. The use of this random
value is to make it somewhat more difficult for the server to be
"accidentally" killed by a user or an attacker.