Bidirectional messages between server and Safari browser.
I've now got the server to the point where it can exchange messages
with the Safari brower. In the test code, a message is send from
client to server over WebSockets and this is simply echoed back (the
browser shows the response text in a div).

Chrome OTOH, is not happy that the HTTP response headers aren't quite
compliant with RFC6455 (I'm not yet sending Sec-WebSocket-Protocol).