Move the UTF code from pstore-json to -support.
The WebSocket spec. requires that text data payloads are validated as
legal UTF-8. Fortunately, we've got code that can do this already in
the JSON parser which converts UTF-8/16 to UTF-32 before re-encoding
it as UTF-8 and passing the text to its consumer.

Move that code to support/utf. This gets all of the Unicode-related
code in one place.