The WS server can now "push" messages. Add an uptime thread demo.
Add a publish-and-subscribe mechanism which can be used to send data
from one part of the program (the publisher) to any number of
recipients (the subscribers). The URI passed as part of the HTTP->WS
upgrade handshake is used to pass the name of the channel to which the
WS server should subscribe.

Add a simple "uptime" thread which acts as a demo for this
behavior. This pushlishes the server up-time every second as a JSON object.