Add a switch to provide a name for the file system object.
Added a --var switch which allows the user to specify the name of the
variable used as the file system objects: previously this was
hardwired to "fs". A namespace may be specified by using double-colon
separators (e.g. "a::b::c": will create a variable "c" inside
namespace "b" which is, in turn, inside namespace "a").