Split the enable and disable warnings code.
Previously, when compiling inside LLVM we saw a significant number of
unwanted warnings from the pstore code because the warning options
were not being set consistently. This change splits the warning
settings into two: a function which sets the base warning state, and a
function which disables the unwanted warnings.