Add an alignment check to getrw<T> and getro<T>.
This upgrades what was previously an assertion to a proper runtime
check. Now that we have type-based addresses and extents it's possible
to also check that the alignment of the type seems reasonable in the
database getXX APIs.

AFL triggered one of these assertions by creating a corrupt input
file: this triggered an assertion rather than a proper error.