CID343562: restrict the types that can be assigned to bit_field.
Coverity was unhappy about a dynamic assertion being used to check the validity of the input type: that can result in an expression whose operands don't affect the result. Instead, I now use enable_if<> to do this so that the restriction is now done at compile-time. This forces the correct types to be used, but that's probably A Good Thing anyway.