Move our make_unique<>() function from namespace std to pstore.
std::make_unique<>() was introduced in C++14 and we compile for
C++11. We provide a version of make_unique() but this was previously
being injected into namespace std which isn't a good idea. Some
platforms (Windows) unconditionally provide make_unique so we need to
play nicely with this.

This change moves our make_unique() implementation to the pstore
namespace. If the host library supports it, then the function is a
trivial wrapper for std::make_unique<>().