Remove a special case from the check() code.
A directory check now explicitly verifies that for a visit to a
previously checked directory. This avoids the need to have a special
case check for the '.' and '..' entries and can handle general
hard-links if that turns out to be useful.

The check for '.' and '..' matching the expected values was broken.

Remove a TODO: just note that we perform a linear search sometimes.

Add some doxygen documentation.