Sanitize the handling of wide/narrow streams and Windows TCHAR.
We previously had many duplicate definitions of the error/output
streams which were wcerr/cwout on Windows and cerr/cout everywhere
else. There was also a rather ah-hoc to handling the Windows _tmain()
command-line argument encoding and the definition of TCHAR on
non-Windows platforms.

This is all now rationalized in the cmd_util/tchar.h/.cpp TU.