internal/linear_node are not trivially copyable so don't use memcpy.
We were using memcpy to copy the body of both linear and internal node
structures from their in-heap version to the store during a
flush. Unfortunately, these types are non "trivially copyable" and
this means that we need to be very careful when using memcpy() or
memmove(). Instead, we now use placement new and a copy constructor to
achieve the same result.

Complete the "rule of 5" for these types.