Fix a bug introduced by ffc6b8db9aea9316c3f592be25082b87499dd42c.
The internal node structure defined its child array at maximum
size. This made its memory-management more straightforward but this
was incompatible with the way that they're held in the store: there we
just record the number of elements that are used.

This makes the ctors private and introduces a set of allocate() memory
functions which allocate and construct the object correctly.