Remove an unnecessary FIXME. More constexpr in address.
Checking shows that the request_spans_regions() function is always
safe because sat_ is over-allocated by one element. Added a note a an
assertion to that effect.

More constexpr in address.

sat_elements is defined in terms of max_segment rather than the raw
number of beats. This should make the relationship between these
values a little more obvious.