Fix a bug in request_spans_regions().
The implementation of request_spans_regions() only checks whether the
request start address and subsequent segment were different. This was
adequate for the vast majority of requests.  However, it was possible
to construct specific scenarios in which this was wrong: see the unit
tests in test_storage.cpp.

Minor tidy-up of some related code (such as using gsl::not_null and
eliminating some single-use variables).