Remove multiple definitions of get_last_error(). Remove a FIXME.
The FIXME was about adding a timeout to the quit connect() call. This
isn't necessary because we are connecting to a port on locahost which
is known to be open and because there is an implicit timeout value in
the connect() implementation.

We had multiple definitions of the (trivial) get_last_error()
function. This is now a single definition in error.hpp.