Some options can now accept a comma-separated list of values.
Extend the --fragments/--compilations/--debug-file-headers so that
they can accept a comma-separated list of values as well as being
specified multiple times as ways of asking for more that one value to
be dumped.

Added a simple function to split a string into comma-separated
parts. This is used from the dump/switch code.