Fixed Windows error 33 whilst opening a store file (issue #40).
If one process held the transaction lock whilst another attempted to
open the same database file, on Windows we would receive an error
33 (ERROR_LOCK_VIOLATION, "The process cannot access the file
because another process has locked a portion of the file.").

This was because we were using some of the bytes of the file header
for both the garbage-collection range lock (used by the GC tool to
know whether another process has the file open) and the transaction
range lock (used to implement the transaction mutex).

During the early stages of opening the file we read the header (using
ReadFile rather than via a memory-mapped file).

The solution adopted is to add a new data block after the file header,
the lock-block. This contains 8 byte fields for both locks and is not
read or written by the store (except during the initial file build
process). This allows is to be safely used for file locks on Windows.