Fix a race when creating database files.
If two processes attempted to create the same database file at the
same time, this may fail; the symptom is different between Windows and
the other OSs but the problem is essentially the same.

Initial opening the database file is split into 4 stages:
1. Try to open the file. If this succeeds, we're done.
2. If the file is not found (and the user's arguments tell us to
create one), we must create a temporary file in the same directory. We
use the same directory to likely ensure that the temporary file is on
the same volume as the destination.
3. Populate the file with an empty pstore shell. (That is, a header,
lock-block, and zeroth transaction footer.)
4. Atomically rename the file to give it its final name.

The final step was racy on both Windows and POSIX hosts. On Windows,
an error was raised if a second process had created the file by the
time we reach step 4; on POSIX, the rename() function overwrites the
destination file if another process had written it by step 4.

Fixes #41 and SNSystems/llvm-project-prepo#44.