Rename httpd->http.
The HTTP library should not have been named "httpd". The 'd' suffix is
usually used to incidate a daemon process whereas the HTTP library is
just that: a library, not a deamon process. This usefully means that
we can drop the library's "-lib" suffix since we no longer have a name
clash with the real server tool (named 'httpd').