Further fixes for VS2019.
It turns out that the cmake generator for VS2019 doesn't support the
"Win64" decoration at the end of its name. This was used in
VS2017 (and possibly others) to indicate that we want to build for
64-bit Windows. Instead, we now need to pass an additional switch "-A
x64".

This has resulted in much more significant changes to the code since
we're dealing with a collection of switches rather than a simple
generator-name string.

Tested with builds on Mac, Linux, Windows+VS2017,
Windows+VS2017+VS2019, Windows+VS2019.