Issue #43 ("Add support for VS2019 to make_build") (#44)
* Issue #43 ("Add support for VS2019 to make_build")

We must discover which versions of Visual Studio are installed then
pick the correct generator string for the latest version that we
find. If an error occurs, we default back to the cmake default generator.

* Minor Fixes.

'installations' must be a list rather than a generator; test options.generator.

* Further fixes for VS2019.

It turns out that the cmake generator for VS2019 doesn't support the
"Win64" decoration at the end of its name. This was used in
VS2017 (and possibly others) to indicate that we want to build for
64-bit Windows. Instead, we now need to pass an additional switch "-A
x64".

This has resulted in much more significant changes to the code since
we're dealing with a collection of switches rather than a simple
generator-name string.

Tested with builds on Mac, Linux, Windows+VS2017,
Windows+VS2017+VS2019, Windows+VS2019.

* A general clean-up of the code.

Here I'm trying to simplify the code to make it as linear as possible.