The lock-test tool should not be installed.
Use add_pstore_executable() rather than add_pstore_tool() to create
the pstore-lock-test target since this is a program used to test the
transaction lock rather than an end-user tool.

(IMO, the distinction between "tool" and "executable" is far too
subtle and the names don't suggest that the difference is that one is
installed an the other is not. I'll change this soon...)