Provide a default value for the visibility_type field.
Commit e62dc469f4bcd6afd31cf35f9a3efdbc99359202 added an additional
argument to the compilation_member ctor. This resulted in two
compilation failures in the rld project.

This change provides a default value for that new argument.

I usually prefer to avoid default arguments, but fell that it's okay
in this case because I'm hopeful that the wholesale import of ELF
"visibility" into the repo won't be permanent and we'll find a more
general way to implement shared-library semantics.