Beginning to think about the size of some of the on-disk structures.
Some of our on-disk structures aren't using space terribly
efficiently. This is particularly relevant for the mcrepo structures
which may each see thousands of instances for a reasonably large
target program.

This small step folds two of the compilation fields into a
byte and reduces the maximum number of compilation members from 2^64
to 2^32 (which should be plenty!). Unfortunately, I can't yet go
further with this struct because the compiler is stuffing host
pointers into one of the other fields.