Enable a number of previously disabled clang warnings.
We had a number of clang's warnings disabled. This was originally
because we were using a revision of Google Test that produced a fair
number of those warnings. Now that we've got a more recent drop of the
test framework, we can re-enable and update to resolve most of them.