Somewhat resolve a TODO in descriptor_condition_variable.
On POSIX, the descriptor_condition_variable is used in a signal
handler which limits the calls that we can safely make. For this
reason, the class previously had no means to report an error writing
to its underlying pipe.

This change splits the API in two: inside the signal handler call
notify_all_no_except(), in all other uses call notify_all().

Switch to using poll() rather than select() in the wait() method.

Correct handling of EAGAIN and EINTR in the write() call.