Correct handling of single letter arguments.
This means that two single letter boolean flags 'a' and 'b' can be set
simply with '-ab'. If a single letter flag 'a' takes an argument it
can be written as '-avalue'.

For the pstore-dump arguments that take a digest, move parsing into a
specialized modue. Ensure that the help shows the argument type as 'digest'.