Fix function casing and remove unused template param.
The original ideal was that the pstore utilities would use the LLVM
command-line parser when building inside LLVM, and a local
implementation when building standalone. The local implementation was
interface-compatible (to the extend required by the features being
used) with the LLVM one.

However, LLVM doesn't handle Windows Unicode support in the same way:
standalone pstore wanted to be a fully fledge "Unicode Aware" program
whereas LLVM uses the MBCS interfaces. This caused extra complexity
and confusion.

Eventually, I decided that being interface-compatible was more trouble
than it was worth. This means that we can begin to diverge from the
LLVM naming convention and template parameters.