Add pstore include files to an inside-LLVM configure.
Working inside LLVM but without the pstore include files in my IDE was
proving to be rather inconvenient. This change adds all of the pstore
include files in both standalone and inside-LLVM builds.

Additionally, it's no longer necessary to provide the full path of
each include file (which resulted in a lot of duplication in the
CMakeLists.txt files).