Add close_noex() for the file_handle dtor. Clean up the win32 code.
Rather than using no_ex_escape() in the file_handle dtor we instead
call a function which will not raise an exception in the event of an
error.

Refactor the error-raising code in the win32 implementation to
eliminate all of the individual code which constructs the message
associated with an error. Eliminate single-use variables.