Fix for issue#55. Store the process handle correctly.
The Windows API defines a HANDLE as void*. Unfortunately, that means
that we don't get much in the way of type safety checks from the
compiler. Here, the bug was that we used the raw process_pair pointer
rather than the process handle, which is one of the members of the
process pair.

Although the code was definitely wrong, the unit test failure was
caused by the WaitForMultipleObjects() function returning an error (it
was not being handed valid HANDLEs) and this resulting in a call to
raise(). In the standalone build, the code recovered from that
error. However, in an "inside LLVM" build, that error caused a call to
std::exit(). This was the cause of the error message from Google Mock.